### 2019-03-28 #1348 @katyhuff
Sorry -- I misread the README. I see now it's a python package mimicking an R package. I am ok with editing this, though the author should be alerted that I will be unresponsive from April 3-9.
