### 2019-03-28 #1348 @danielskatz
@katyhuff - this is python, as far as I can tell, though it is based on something in R
