### 2019-03-28 #1348 @kmichael08
Hello, yes it's a Python package. There is just a mention of the same idea being implemented earlier in R
(by one of the co-authors).
