### 2019-04-16 #1389 @whedon
[REVIEW]: pyCeterisParibus: explaining Machine Learning models with Ceteris Paribus Profiles in Python
**Submitting author:** @kmichael08 (<a href="http://orcid.org/0000-0002-9181-0126">Michał Kuźba</a>)
**Repository:** <a href="https://github.com/ModelOriented/pyCeterisParibus" target ="_blank">https://github.com/ModelOriented/pyCeterisParibus</a>
**Version:** v0.5.2
