### 2019-04-17 ModelOriented/pyCeterisParibus#23 @justinshenk
Scrollbars block data in plots
**Describe the bug**
Scroll bars block the data in plots:

**To Reproduce**
Run code from README tutorial.

**Expected behavior**
Expected to be able to read data behind scroll bar, but can't.

**Screenshots**
![image](https://user-images.githubusercontent.com/10270308/56275400-ec11e200-6100-11e9-8b10-010e3a153135.png)

Mac OSX, Chrome
