### 2019-04-17 #1389 @justinshenk
**Overview:** pyCeterisParibus is a library for explaining machine learning models with _ceteris paribus_ profiles. These are useful for adding to visual story telling and supporting model interpretability. The idea is great, the implementation is clean, and I may use it in some of my projects. Some minor improvements are suggested below.

**Installation:** Installed without issues via pip and local copy of the source code.

**Functionality:** I was able to run the example on my Mac, but was not able to load the plot, due to an issue with how file paths are handled on Macs. I have opened a pull request at https://github.com/ModelOriented/pyCeterisParibus/pull/24 fixing this issue on my machine. After this is accepted or otherwise addressed I will consider it completed.
I opened an issue (https://github.com/ModelOriented/pyCeterisParibus/issues/23) regarding the scrollbars obscuring the data. This could be fixed by adding additional padding to the bottom of the frame.

**Performance:** No measure of performance is given, but the model loaded fast on the Titanic dataset.

**Documentation:** The explanation for how the model works could be improved. For example, in the paper the author's write "For this purpose, methods for sampling and selecting neighbouring observations are implemented along with the *Gower's distance [@gower]* function. A more detailed description might be found in the package [documentation](https://pyceterisparibus.readthedocs.io/)." I was not able to find description of Gower's distance in the linked to readthedocs. Adding details of how the model works would be helpful for people who are not familiar with Gower's distance or how it applies to machine learning models.

**Software Paper:** The software paper has a few minor typos or questionable stylistic choices for an academic paper:
- "He died on [the] Titanic"
- "1. [rather first or 1st] class"
- **BUT**. Not really a typo, but stylistically questionable to have a period after "BUT". Would be more fitting of an academic journal without the word.
- Local [I]nterpretable [M]odel-agnostic [E]xplanations (LIME)
- Capitalization of "machine learning"
- "(e.g.[,] a bank)"
- "black-boxes" [no space needed]
- "leads to the next industrial revolution" - unverified claim

**Example Usage:** The notebooks and example scripts without problems.

**References:** Every reference mentioned in the paper is documented as BibTex entries.
