### 2019-04-17 8c41162
correct typos and style in the paper
