### 2019-04-17 ModelOriented/pyCeterisParibus#23 @kmichael08
@flaminka Can you have a look at it?
