### 2019-04-17 d9dfdc4
Describe gower's distance in documentation
