### 2019-04-19 ModelOriented/pyCeterisParibus#23 @flaminka
Hi, thank you for reporting this! :) Unfortunately, I cannot reproduce this error, I've evoked this code on Windows 10  and also on Mac Mojave, in Chrome and everything seems to work. Could you please define which version of Chrome you use? And, if possible, make a screen recording when you evoke this code / make printscreens of Chrome console and Python console?
