### 2019-04-23 ModelOriented/pyCeterisParibus#23 @justinshenk
This issue happens on Chrome 73.0. It is fixed by opening and closing the development tools window. Closing this issue since it seems specific to my computer.
