### 2019-04-26 #1389 @justinshenk
@kmichael08 Thanks for the quick response and edits.

@katyhuff My review (https://github.com/openjournals/joss-reviews/issues/1389#issuecomment-484024162) is now complete.
