### 2019-05-02 #1389 @janfreyberg
## Installation

Installing the package in a fresh docker alpine image leads to dependencies being installed that I don't think need to be, for example sphinx, m2r, codecov, etc.

There are a few ways around this so I haven't made a PR but I think you can do the following:

Split requirements into actual requirements (what's needed to run the package), documentation requirements, and test requirements (e.g. pytest). You can add these as additional requirements using the [extras_require](https://setuptools.readthedocs.io/en/latest/setuptools.html) key in setup.py, or simply install them from txt files wherever you need them.

Additionally, as far as I can tell tensorflow is never imported and so should be removed from the requirements.

I would even go so far as to say XGBoost and sklearn should not be in the requirements, even though you use it in the paper and documentation, becuase it's not essential to the functioning of the package. Instead, you could make a note that people should install them to run the examples.

Otherwise, installation works great.

## Functionality / Performance

This all worked great for me.

## Documentation

I think the docs can be improved:

- the index page should be clearer; I would remove the automated sphinx content and add more of a "landing" page that contains an introduction
- I would consider adding the jupyter notebooks to the docs using a tool like sphinx-jupyter

But that's just a recommendation.

## Paper

The paper is very good. Only point: the R package CeterisParibus is not included in the references.
