+++ paper/paper.bib
@@ -25,6 +25,15 @@
   year={2018}
 }
 
+@misc{CeterisParibus,
+  author = {Biecek, Przemysław},
+  title = {Ceteris Paribus Plots (What-If plots) for explanations of a single observation},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/pbiecek/ceterisParibus}},
+  commit = {5eb44b93a52ee3f9c84dc455e8fcc6a81de4c6e7}
+}
 
 @incollection{SHAP,
 title = {A Unified Approach to Interpreting Model Predictions},
