### 2019-05-03 64ca4c8
reduce essential requirements list
