+++ requirements.txt
@@ -1,16 +1,3 @@
-codecov>=2.0.15
-coverage>=4.5.2
 Flask>=1.0.2
-Keras>=2.2.4
-matplotlib>=3.0.1
-m2r==0.2.1
 numpy>=1.15.4
 pandas>=0.23.4
-pytest>=4.0.1
-pytest-cov>=2.6.0
-scikit-learn~=0.20.0
-scipy>=1.1.0
-sklearn==0.0
-Sphinx>=1.8.3
-tensorflow>=1.12.0
-xgboost>=0.82
