### 2019-05-03 #1389 @kmichael08
Thanks a lot @janfreyberg and @katyhuff!

* I added the citation to the R package
* You're right about the requirements. I kept only essential dependencies in `requirements.txt` and moved others to `requirements-dev.txt` if that's ok. As for the tensorflow, it is required although not directly imported. There is a test using keras, and this needs some backend DL library (tensorflow here).

So, as far as this two sounds ok for you, we can move on. I'll definitely enhance the docs soon and use sphinx-jupyter. Thanks for that!
