### 2019-05-03 #1389 @katyhuff
I'm going through some of the final checks (first up, the bibliography):

- [ ] **Should the following doi** [doi:10.1214/aos/1013203451](https://doi.org/10.1214/aos/1013203451) **be added to:** Friedman, J. H. (2000). Greedy function approximation: A gradient boosting machine. Annals of Statistics, 29, 1189–1232. ? It may be pointing to a different version of the paper than the one you're trying to cite.. but it seems like the right one to me.

- [ ] The link in the Rutz, J (2018) citation is too long and overflows the margins. That's not really your fault -- Our latex template already has `hypersetup[breaklinks..]` so it should be breaking the URL. I don't know why the URL isn't breaking.
@arfon -- any ideas on this URL breaking issue? Perhaps the latex template should use `breaklinks=true` rather than just `breaklinks`?

- bib: https://github.com/ModelOriented/pyCeterisParibus/blob/1f75844d31ec7c5b1442401f1f596b90c657f430/paper/paper.bib#L16
- joss latex template: https://github.com/openjournals/whedon/blob/033eedaceb8080759917955949bbd7db36fcfda6/resources/latex.template#L10
