### 2019-05-03 #1389 @arfon
Weird. If I change the bib file field to:

```
howpublished = {\url{https://www.openrightsgroup.org/blog/2018/machine-learning-and-the-right-to-explanation-in-gdpr}},
```

Then it seems to compile OK. Changing the flag to `breaklinks=true` doesn't seem to fix anything.
