+++ paper/paper.bib
@@ -13,7 +13,7 @@
   title={Machine learning and the right to explanation in GDPR},
   author={Javier Ruiz},
   year={2018},
-  url={https://www.openrightsgroup.org/blog/2018/machine-learning-and-the-right-to-explanation-in-gdpr}
+  howpublished = {\url{https://www.openrightsgroup.org/blog/2018/machine-learning-and-the-right-to-explanation-in-gdpr}}
 }
 
 @article{DALEX,
@@ -70,13 +70,19 @@ url = {http://papers.nips.cc/paper/7062-a-unified-approach-to-interpreting-model
     year = {2015},
   }
 
-@ARTICLE{Friedman00greedyfunction,
-    author = {Jerome H. Friedman},
-    title = {Greedy Function Approximation: A Gradient Boosting Machine},
-    journal = {Annals of Statistics},
-    year = {2000},
-    volume = {29},
-    pages = {1189--1232}
+@article{friedman2001,
+author = "Friedman, Jerome H.",
+doi = "10.1214/aos/1013203451",
+fjournal = "The Annals of Statistics",
+journal = "Ann. Statist.",
+month = "10",
+number = "5",
+pages = "1189--1232",
+publisher = "The Institute of Mathematical Statistics",
+title = "Greedy function approximation: A gradient boosting machine.",
+url = "https://doi.org/10.1214/aos/1013203451",
+volume = "29",
+year = "2001"
 }
 
 @misc{pramit_choudhary_2018_1198885,
