### 2019-05-03 #1389 @katyhuff
Thanks @arfon for the workaround and thank you to @justinshenk @janfreyberg for your excellent reviews.

Thank you @kmichael08 for a strong submission and for engaging actively in the review process! I have looked over the paper, double checked all the DOI links, and have conducted a high level review of the code itself. Everything looks ship-shape to me.

@kmichael08 At this point, please double check the paper yourself, if you want to update your code version (e.g. change v5.0 to some minor release representing today's version) review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission! Until then, now is your moment for final touchups!
