+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages
 PACKAGE_NAME = 'pyCeterisParibus'
 REQUIREMENTS_PATH = 'requirements.txt'
 README_PATH = 'README.md'
-VERSION = '0.5.1'
+VERSION = '0.5.2'
 
 
 def get_requirements():
