### 2019-05-06 #1389 @kmichael08
@katyhuff I updated repository to the v0.5.2 and archived it in Zenodo.
DOI: 10.5281/zenodo.2667756
