### 2019-05-06 #1389 @arfon
> Wait a couple of minutes to verify that the paper DOI resolves https://doi.org/10.21105/joss.01389

Also, please note, Crossref is still having some issues so this DOI doesn't resolve yet.
