### 2019-05-06 #1389 @arfon
> So (@arfon @kyleniemeyer ) do we just run accept again?

There's no need to because the version isn't captured anywhere other than here. The archive DOI is correct right? (This is linked to in the paper)
