### 2019-05-07 #1389 @arfon
> Also, please note, Crossref is still having some issues so this DOI doesn't resolve yet.

Just to let you know, the DOI is now resolving properly.
